Title: 숫자 합치기, Time: 3082ms, Memory: 36MB, Status: Time Limit Exceed - CodeTree

+++ 241116/숫자 합치기/merge-numbers.java
@@ -17,7 +17,7 @@ public class Main {
 			int n = sc.nextInt();
 			list.add(n);
 		}
-		int totalCost = 0;
+		long totalCost = 0;
 		while(list.size()>= 2) {
 			Collections.sort(list);
 			int cost = list.get(0)+list.get(1);
